+++ grid2op/Environment/BaseEnv.py
@@ -621,6 +621,8 @@ class BaseEnv(GridObjects, RandomObject, ABC):
         new_obj._kwargs_attention_budget = copy.deepcopy(self._kwargs_attention_budget)
 
         new_obj._last_obs = self._last_obs.copy()
+        
+        new_obj._has_just_been_seeded = self._has_just_been_seeded
 
     def get_path_env(self):
         """